+++ src/Spin.cpp
@@ -1,5 +1,5 @@
 // See Spin.h for description.
-// Seto Balian, November 29, 2013
+// Seto Balian, December 2, 2013
 
 #include "Spin.h"
 
@@ -8,8 +8,8 @@ unsigned int Spin::calculate_multiplicity() const
   return static_cast<unsigned int> (2.0*Spin::get_quantum_number() + 1.0);
 }
 
-Spin::Spin(const double quantum_number, const double gyromagnetic_ratio,
-           const unsigned int multiplicity) :
+Spin::Spin(const double quantum_number, const double gyromagnetic_ratio, 
+        const unsigned int multiplicity) :
   quantum_number_(quantum_number),
   gyromagnetic_ratio_(gyromagnetic_ratio),
   multiplicity_(multiplicity)
@@ -22,8 +22,8 @@ Spin::Spin() :
   multiplicity_(0.0)
 {/**/}
 
-Spin::Spin(const double quantum_number,
-           const double gyromagnetic_ratio) :
+Spin::Spin(const double quantum_number, 
+        const double gyromagnetic_ratio) :
   quantum_number_(quantum_number),
   gyromagnetic_ratio_(gyromagnetic_ratio),
   multiplicity_(calculate_multiplicity())
@@ -39,7 +39,7 @@ double Spin::get_quantum_number() const
   return quantum_number_;
 }
 
-double Spin::get_multiplicity() const
+unsigned int Spin::get_multiplicity() const
 {
   return multiplicity_;
 }
